+++ asgi_lua/__init__.py
@@ -9,6 +9,7 @@ from .middleware import LuaConfig, LuaMiddleware
 from .presets import McpPolicyOptions, copy_builtin_preset, generate_mcp_preset, list_builtin_presets
 from .promotion import compare_decisions, diff_policies
 from .proxy import ReverseProxyApp, create_proxy_application, run_proxy
+from .quickstart import create_mcp_quickstart
 from .recorder import append_record, load_record_log, record_audit_event, record_policy_request, replay_record_log
 from .redaction import RedactionConfig, append_audit_event, build_audit_event, redact_secrets
 from .runtime import LuaDecisionError, LuaDecisionTrace, LuaRuntimeError
@@ -43,6 +44,7 @@ __all__ = [
     "build_audit_event",
     "compare_decisions",
     "copy_builtin_preset",
+    "create_mcp_quickstart",
     "create_proxy_application",
     "diff_policies",
     "generate_mcp_preset",